fixed dependency conflicts deleted index.js and added its contents to main.js add auth folder to .gitignore
