+++ package.json
@@ -7,11 +7,11 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "electron": "^latest",
+    "axios": "^1.7.2",
+    "electron": "^31.1.0",
     "express": "^4.17.2",
-    "axios": "^0.24.0",
-    "twitter": "^1.7.1",
-    "facebook-node-sdk": "^0.8.0"
+    "facebook-node-sdk": "^0.2.0",
+    "twitter": "^1.1.0"
   },
   "scripts": {
     "start": "electron ."
